Rename NotifierCommand to Notifier
