+++ src/ScrumMaster/IO/NotifierOutput.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Chemaclass\ScrumMaster\Command;
+namespace Chemaclass\ScrumMaster\IO;
 
 use Chemaclass\ScrumMaster\Channel\ChannelResult;
 use Chemaclass\ScrumMaster\Channel\ReadModel\ChannelIssue;
-use Chemaclass\ScrumMaster\Command\IO\OutputInterface;
+use Chemaclass\ScrumMaster\IO\OutputInterface;
 use Chemaclass\ScrumMaster\Common\Request;
 
 final class NotifierOutput
