+++ tests/ScrumMasterTest/Unit/NotifierTest.php
@@ -7,22 +7,22 @@ namespace Chemaclass\ScrumMasterTests\Unit\Command;
 use Chemaclass\ScrumMaster\Channel\ChannelInterface;
 use Chemaclass\ScrumMaster\Channel\ChannelResult;
 use Chemaclass\ScrumMaster\Channel\ReadModel\ChannelIssue;
-use Chemaclass\ScrumMaster\Command\NotifierCommand;
-use Chemaclass\ScrumMaster\Command\NotifierInput;
+use Chemaclass\ScrumMaster\Notifier;
+use Chemaclass\ScrumMaster\IO\NotifierInput;
 use Chemaclass\ScrumMaster\Jira\JiraHttpClient;
 use Chemaclass\ScrumMasterTests\Unit\Concerns\JiraApiResource;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-final class NotifierCommandTest extends TestCase
+final class NotifierTest extends TestCase
 {
     use JiraApiResource;
 
     /** @test */
     public function zeroNotificationsWereSent(): void
     {
-        $command = $this->notifierCommandWithChannelIssues([]);
-        $result = $command->execute($this->notifierInput());
+        $notifier = $this->notifierCommandWithChannelIssues([]);
+        $result = $notifier($this->notifierInput());
         /** @var ChannelResult $channelResult */
         $channelResult = reset($result);
         $this->assertEmpty($channelResult->channelIssues());
@@ -36,8 +36,8 @@ final class NotifierCommandTest extends TestCase
             'KEY-2' => ChannelIssue::withCodeAndAssignee(200, 'jira.user.2'),
         ];
 
-        $command = $this->notifierCommandWithChannelIssues($issues);
-        $result = $command->execute($this->notifierInput());
+        $notifier = $this->notifierCommandWithChannelIssues($issues);
+        $result = $notifier($this->notifierInput());
         /** @var ChannelResult $channelResult */
         $channelResult = reset($result);
         $this->assertEquals($issues, $channelResult->channelIssues());
@@ -48,13 +48,13 @@ final class NotifierCommandTest extends TestCase
         return NotifierInput::new('company.name', 'Jira project name', ['status' => 1]);
     }
 
-    private function notifierCommandWithChannelIssues(array $channelIssues): NotifierCommand
+    private function notifierCommandWithChannelIssues(array $channelIssues): Notifier
     {
         /** @var ChannelInterface|MockObject $channel */
         $channel = $this->createMock(ChannelInterface::class);
         $channel->method('sendNotifications')->willReturn(ChannelResult::withIssues($channelIssues));
 
-        return new NotifierCommand(
+        return new Notifier(
             new JiraHttpClient($this->mockJiraClient([])),
             [$channel]
         );
